requirements.txt
Flask==1.1.1
google-cloud-automl